[Updated the Balance Ball to use Scriptable Object Brains] (#903)
Make sure you have ENABLE_TENSORFLOW in the project settings
Look into the Academy and notice there is now a training hub instead of Child brains
10 platforms use the internal brain, one the player brain and one a random decision
External brain is no longer present. Inf you want to train a brain, check train in the training hub
Look into the Brains folder in the 3DBall folder to look at the brain scriptable objects
What now ? Test it out, see if brains as scriptable objects make sense and if we want to implement it into develop.