Check that worker port is available in RpcCommunicator
Previously the RpcCommunicator did not check the port or create the
RPC server until `initialize()` was called.  Since "initialize"
requires the environment to be available, this means we might create
a new environment which connects to an existing RPC server running
in another process.  This causes both training runs to fail.

As a remedy to this issue, this commit moves the server creation into
the RpcCommunicator constructor and adds an explicit socket binding
check to the requested port.