Limit pytest <4.0.0
Some of our tests fail with the new pytest v4.0.x versions because of
a change to how fixtures work. Since these versions are the latest
release in PyPI now, unit tests will fail on a fresh install.

This change limits the version range in setup.py to fix the issue
until we can make a decision about upgrading.