Support both 32-bit and 64-bit types in UnityEnvironment (#1471)
We check for the single brain case in UnityEnvironment by checking
for applicable non-dict types in the step arguments. However for ints
and floats we just use `np.int_` and `np.float_` for the check, which
are the defaults for your system.

This means if you are using an application (like baselines in #1448)
which uses the wrong int/float size an error will be thrown.  This
change explicitly allows both 32 and 64-bit numbers.