Fix for divide-by-zero error with Discrete Actions  (#1520)
* Enable buffer padding to be set other than 0

Allows buffer padding in AgentBufferField to be set to a custom value. In particular, 0-padding for `action_masks` causes a divide-by-zero error, and should be padded with 1’s instead.

This is done as a parameter passed to the `append` method, so that the pad value can be set right after the instantiation of an AgentBufferField.