Fix '--slow' flag after environment updates
A change was made to the way the "train_mode" flag was used by
environments when SubprocessUnityEnvironment was added which was
intended to be part of a separate change set.  This broke the CLI
'--slow' flag.  This change undoes those changes, so that the slow
/ fast simulation option works correctly.

As a minor additional change, the remaining tests from top level
'tests' folders have been moved into the new test folders.