Remove "external_brains" arg for TrainerController
TrainerController depended on an external_brains dictionary with
brain params in its constructor but only used it in a single function
call.  The same function call (start_learning) takes the environment
as an argument, which is the source of the external_brains.

This change removes the dependency of TrainerController on external
brains and removes the two class members related to external_brains
and retrieves the brains directly from the environment.