Add timeout for communicator exchange
When we initially connect to the environment using RPCCommunicator,
the connection is polled so we don't hang forever on `.recv()` when
the environment wasn't launched or failed.  However we don't currently
have any similar check for the exchanges mid-training-run.

This change applies the same timeout from initialization to each exchange,
and extends the default `timeout_wait` to 60 seconds to generally improve
the chances we won't have a mismatch between environment launch time and
the trainer timeout.

Tested on: single-env and multi-env cases.  Killed 1 environment process
manually and saw that the model was saved appropriately and all processes
closed.