Add command-line flag for example env scene selection
The Startup scene used for us to retrain and bundle example scenes
together already accepts the SCENE_NAME environment variable to choose
a scene.  This change adds support for command-line selection of the
scene name so that `--env-args` can be used to choose the scene.

If both the environment variable and command-line flag are provided, the
scene will use the environment variable.