Shorten timeout duration for environment close
The timeout duration for closing an environment was set to the
same duration as the timeout when waiting for a response from the
still-running environment.  This led to long waits for the error
response when communication version wasn't matching.

This change updates the timeout duration so that we wait a fixed
timeout of 5 seconds before force-killing the environment worker.