Remove "docker target" feature (#3687)
The "docker target" feature and associated command-line flag
--docker-target-name were created for use with the now-deprecated
Docker setup.  This feature redirects the paths used by learn.py
for the environment and config files to be based from a directory
other than the current working directory.  Additionally it wrapped
the environment execution with xvfb-run.

This commit removes the "docker target" feature because:

* Renaming the paths doesn't fix any problem.  Absolute paths can
  already be passed for configs and environment executables.
* Use of xserver, Xvfb, or xvfb-run are independent of mlagents-learn
  and can be used outside of the mlagents-learn call.  Further, xvfb-run
  is not the only solution for software rendering.