Remove unused TouchedTarget() methods from examples
Crawler, Walker, and Worm all have TouchedTarget methods to add +1 reward
that are, as near as I can tell, completely unused. These are not mentioned
anywhere in documentation and are potentially very confusing to someone
examining the example code. Remove this methods to make the code cleaner
and easier to read.