Move RigidBodySensorComponent update back to end and expand RootBody null check
Having the update of the internal object where it was was causing problems,
so put it back at the end. Also, for some reason I will never understand,
the `ReferenceEquals` check on `RootBody` was sometimes returning false
even when it was set to `null`, so add an extra check for Unity `null`.